Add `SwiftSDKBundleStore`, fix tests
